+++ slsSupportLib/include/sls/string_utils.h
@@ -60,6 +60,6 @@ bool is_int(const std::string &s);
 bool replace_first(std::string *s, const std::string &substr,
                    const std::string &repl);
 
-std::pair<std::string, int> ParseHostPort(const std::string &s);
+std::pair<std::string, uint16_t> ParseHostPort(const std::string &s);
 
 } // namespace sls
